requirements.txt
# StockTrader Bot - Modern Full-Stack Architecture Requirements
# Python 3.12+ 

# ===== FASTAPI BACKEND DEPENDENCIES =====
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
websockets>=12.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# ===== DATABASE (OPTIONAL) =====
sqlalchemy>=2.0.0
alembic>=1.12.0

# ===== CORE DATA PROCESSING =====
pandas>=2.1.0
numpy>=1.24.0,<2.0  # Pin to < 2.0 for pandas_ta compatibility
yfinance>=0.2.18

# ===== TECHNICAL ANALYSIS =====
pandas-ta>=0.3.14b
# talib-binary>=0.4.19  # For additional TA indicators (Not compatible with Python 3.12 yet)

# ===== MACHINE LEARNING =====
scikit-learn>=1.3.0
torch>=2.0.0
torchvision>=0.15.0
joblib>=1.3.0

# ===== DATA VISUALIZATION (FOR BACKEND PROCESSING) =====
plotly>=5.17.0
matplotlib>=3.7.0

# ===== HTTP REQUESTS =====
requests>=2.31.0
requests-oauthlib>=1.3.1

# ===== SECURITY & AUTHENTICATION =====
cryptography>=41.0.0
bcrypt>=4.0.1

# ===== NOTIFICATIONS =====
twilio>=8.10.0

# ===== CONFIGURATION & ENVIRONMENT =====
python-dotenv>=1.0.0
pyyaml>=6.0

# ===== LOGGING & MONITORING =====
structlog>=23.1.0

# ===== DEVELOPMENT & TESTING =====
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.7.0
ruff>=0.0.287
mypy>=1.5.0

# ===== LEGACY STREAMLIT (FOR GRADUAL MIGRATION) =====
# Note: These will be removed as we migrate to Next.js frontend
streamlit>=1.28.0
streamlit-authenticator>=0.2.3

# ===== UTILITIES =====
# pathlib2>=2.3.7  # Not needed in Python 3.12+ (pathlib is built-in)
typing-extensions>=4.7.0
python-dateutil>=2.8.0
pytz>=2023.3

# ===== E*TRADE API INTEGRATION =====
# Note: E*TRADE uses OAuth1 which is covered by requests-oauthlib above

# ===== ASYNC SUPPORT =====
# asyncio>=3.4.3  # Built-in to Python 3.12+
aiofiles>=23.1.0

# ===== JSON & DATA SERIALIZATION =====
orjson>=3.9.0  # Fast JSON serialization for APIs

# ===== CORS & MIDDLEWARE =====
# Included in FastAPI

# ===== FILE OPERATIONS =====
# Standard library pathlib is sufficient, but for advanced operations:
watchdog>=3.0.0  # For file monitoring

# ===== PERFORMANCE MONITORING =====
psutil>=5.9.0

# ===== TIMEZONE HANDLING =====
# pytz already included above

# ===== DEVELOPMENT TOOLS =====
pre-commit>=3.4.0

# ===== COMPATIBILITY FIXES =====
# Pin setuptools to suppress pandas_ta pkg_resources deprecation warnings
setuptools<81.0.0
